+++ lib/exqlite/sqlite3.ex
@@ -169,6 +169,13 @@ defmodule Exqlite.Sqlite3 do
   @doc """
   Once finished with the prepared statement, call this to release the underlying
   resources.
+
+  This should be called whenever you are done operating with the prepared statement. If
+  the system has a high load the garbage collector may not clean up the prepared
+  statements in a timely manner and causing higher than normal levels of memory
+  pressure.
+
+  If you are operating on limited memory capacity systems, definitely call this.
   """
   @spec release(db(), statement()) :: :ok | {:error, reason()}
   def release(conn, statement) do